moved the split direction into hierarchical manips
